separation of ruby code into multiple files
