+++ lib/splooshed/food_line.rb
@@ -1,0 +1,5 @@
+class FoodLine
+  def self.parse amount, unit, food
+    RecipeLine.parse "#{amount} #{unit} of #{food}"
+  end
+end
