+++ lib/splooshed/recipe.rb
@@ -1,0 +1,5 @@
+class Recipe
+  def self.parse lines
+    lines.split("\n").map {|l| RecipeLine.parse l }
+  end
+end
